1508 : changes based on comments
